Support building against the www.metissian.com Mac OS X kits
Update svn_cmd.py to allow test that assume old svn behavior to see that
behavior when using newer svn versions.


git-svn-id: https://svn.code.sf.net/p/pysvn/code/trunk/pysvn/Extension@513 8a6430f3-a0d1-4a6a-aeff-73e05446c159

+++ Source/pysvn_callbacks.cpp
@@ -157,8 +157,9 @@ void pysvn_context::contextProgress
     Py::Callable callback( m_pyfn_Progress );
 
     Py::Tuple args( 2 );
-    args[0] = Py::Int( progress );
-    args[1] = Py::Int( total );
+    // on some platforms apr_off_t is int64
+    args[0] = Py::Int( static_cast<long int>( progress ) );
+    args[1] = Py::Int( static_cast<long int>( total ) );
 
     Py::Object results;
